[v0] Eliminate v0 protocol as it's deprecated
